requirements.txt
# Core Dependencies
langchain-core==0.3.43
langchain-openai==0.3.8
langgraph==0.3.5
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.1

# Azure OpenAI
azure-identity==1.16.0
azure-keyvault-secrets==4.7.0

# Additional Utilities
pydantic==2.6.4
tinydb==4.8.2
requests==2.32.3
orjson==3.10.15
tiktoken==0.9.0

# Testing (Optional)
pytest==8.2.1
pytest-asyncio==0.23.5